Conflicts:
	towercraftdefense/src/towercraftdefense/bo/Entite.java
	towercraftdefense/src/towercraftdefense/bo/Zone.java
	towercraftdefense/src/towercraftdefense/jeu/Partie.java
	towercraftdefense/src/towercraftdefense/views/Fenetre.java

+++ towercraftdefense/src/towercraftdefense/bo/biosphere/Monstre.java
@@ -7,7 +7,7 @@ import java.awt.Image;
  */
 public class Monstre extends Personnage {
 
-    public Monstre(double coordx, double coordy, double width, double height, double vitesse, double angle, int health, int damage) {
+    public Monstre(int coordx, int coordy, int width, int height, int vitesse, int angle, int health, int damage) {
         super(coordx, coordy, width, height, vitesse, angle, health, damage);
     }
