+++ towercraftdefense/src/towercraftdefense/bo/biosphere/Personnage.java
@@ -11,12 +11,12 @@ public abstract class Personnage extends EntiteMobile {
     private int health;
     private int damage;
   
-    public Personnage(double coordx, double coordy, double width, double height, double vitesse, double angle, int health) {
+    public Personnage(int coordx, int coordy, int width, int height, int vitesse, int angle, int health) {
         super(coordx, coordy, width, height, vitesse, angle);
         this.health = health;
     }
 
-    public Personnage(double coordx, double coordy, double width, double height, double vitesse, double angle, int health, int damage ) {
+    public Personnage(int coordx, int coordy, int width, int height, int vitesse, int angle, int health, int damage) {
         super(coordx, coordy, width, height, vitesse, angle);
         this.health = health;
         this.damage = damage;
